removal of nested Repository classes
