+++ Arius/Arius.Core/Commands/Restore/RestoreCommand.cs
@@ -89,7 +89,7 @@ internal class RestoreCommand : ICommand<IRestoreCommandOptions>
         if (!chunkRehydrating)
         {
             logger.LogInformation("All binaries restored, deleting temporary hydration folder, if applicable");
-            await repo.Chunks.DeleteHydrateFolderAsync();
+            await repo.DeleteHydrateFolderAsync();
         }
         else
         {
